+++ CHANGELOG.md
@@ -58,7 +58,7 @@
 - Enh #917: Rename `ColumnSchemaInterface` to `ColumnInterface` (@Tigrov)
 - Enh #919: Replace `name()` with immutable `withName()` method in `ColumnInterface` interface (@Tigrov)
 - Enh #921: Move `DataType` class to `Yiisoft\Db\Constant` namespace (@Tigrov)
-- Enh #926: Refactor `DbArrayHelper` (@Tigrov)
+- Enh #926, #954: Refactor `DbArrayHelper` (@Tigrov)
 - Enh #920: Move index constants to the appropriate DBMS driver's `IndexType` and `IndexMethod` classes (@Tigrov)
 - New #928: Add `ReferentialAction` class with constants of possible values of referential actions (@Tigrov)
 - Enh #929: Refactor array, structured and JSON column type expressions and expression builders (@Tigrov)
@@ -76,6 +76,8 @@
 - Enh #943: Add `getCacheKey()` and `getCacheTag()` methods to `AbstractPdoSchema` class (@Tigrov)
 - Enh #944: Added `setWhere()` as method a forced for overwriting `where()` (@lav45)
 - Enh #925: Add callback to `Query::all()` and `Query::one()` methods (@Tigrov)
+- Enh #925, #951: Add callback to `Query::all()` and `Query::one()` methods (@Tigrov, @vjik)
+- New #954: Add `DbArrayHelper::arrange()` method (@Tigrov)
 
 ## 1.3.0 March 21, 2024